+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/schedule/FrontScheduleEntity.java
@@ -32,11 +32,11 @@ public class FrontScheduleEntity extends NewCommonMappedClass {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "idx")
-    private Integer idx;
+    private Long idx;
 
     @Column(name = "model_idx")
     @ApiModelProperty(value = "모델 idx", required = true)
-    private Integer modelIdx;
+    private Long modelIdx;
 
     @Column(name = "model_schedule")
     @Lob
