+++ src/main/java/com/tsp/new_tsp_front/api/model/service/FrontModelJpaApiService.java
@@ -136,7 +136,7 @@ public class FrontModelJpaApiService {
      * </pre>
      */
     @Transactional(readOnly = true)
-    public Integer favoriteModelCount(Integer idx) throws TspException {
+    public Integer favoriteModelCount(Long idx) throws TspException {
         try {
             return this.frontModelJpaRepository.favoriteModelCount(idx);
         } catch (Exception e) {
